remove utilities.py from valor_lite.text_generation
